refactor: separate classes for file validation
